Реорганизовал код reviews_service: вынес функции работы с MongoDB в mongo_service.py, добавил маршруты в routes.py и схемы Pydantic в schemas.py. Упрощён main.py, теперь использует router для обработки запросов.
