+++ reviews_service/app/mongo_service.py
@@ -1,0 +1,7 @@
+from motor.motor_asyncio import AsyncIOMotorClient
+
+from os import getenv
+
+client = AsyncIOMotorClient(getenv("MONGODB_URL"))
+db = client.reviews_db
+reviews_collection = db.reviews
